requirements.txt
pytak==7.1.0
python-decouple==3.8
